# docs/api/v1-openapi.yaml
openapi: 3.1.0
info:
  title: GotTheKeys JSON API
  version: "1.0.0"
  summary: Buyer/renter API for GotTheKeys, consumed first by the iOS app.
  description: |
    See `docs/api/v1-spec.md` for the design rationale. All endpoints under
    `/api/v1`. Authentication is JWT bearer; most endpoints require a token,
    catalogue + reference endpoints are public.
  contact:
    name: GotTheKeys
servers:
  - url: https://gotthekeys.example.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/properties:
    get:
      tags: [Catalogue]
      summary: List published properties
      security: []
      parameters:
        - { in: query, name: q,             schema: { type: string } }
        - { in: query, name: sale_status,   schema: { type: string, enum: [for_sale, for_rent] } }
        - { in: query, name: town_city,     schema: { type: string } }
        - { in: query, name: min_bedrooms,  schema: { type: integer } }
        - { in: query, name: min_price,     schema: { type: integer, description: "in pence" } }
        - { in: query, name: max_price,     schema: { type: integer, description: "in pence" } }
        - { in: query, name: property_type, schema: { type: string, enum: [House, Flat] } }
        - { in: query, name: featured,      schema: { type: boolean } }
        - { in: query, name: sort,          schema: { type: string, enum: [recommended, newest, price_asc, price_desc] } }
        - { in: query, name: page,          schema: { type: integer, minimum: 1 } }
        - { in: query, name: per_page,      schema: { type: integer, minimum: 1, maximum: 50 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PropertyCollection" }
  /api/v1/properties/{id}:
    get:
      tags: [Catalogue]
      summary: Property detail
      security: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PropertyDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
        "410": { $ref: "#/components/responses/Gone" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Gone:
      description: Resource has been withdrawn
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:    { type: string }
            message: { type: string }
            request_id: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field:   { type: string }
                  code:    { type: string }
                  message: { type: string }
...
    PropertySummary:
      type: object
      properties:
        id:                  { type: integer }
        listing_state:       { type: string }
        sale_status:         { type: string, enum: [for_sale, for_rent] }
        property_type:       { type: string }
        tagline:             { type: string }
        bedrooms:            { type: integer }
        bathrooms:           { type: integer }
        asking_price_pence:  { type: integer }
        asking_price_display:{ type: string }
        currency:            { type: string }
        featured:            { type: boolean }
        primary_photo:
          type: object
          nullable: true
          properties:
            url:     { type: string }
            caption: { type: string, nullable: true }
        saved_by_me:         { type: boolean, nullable: true }
        next_available_slot: { type: string, format: date-time, nullable: true }
        url:                 { type: string }
        address:             { $ref: "#/components/schemas/Address" }
    PropertyDetail:
      allOf:
        - { $ref: "#/components/schemas/PropertySummary" }
        - type: object
          properties:
            description:                 { type: string }
            tenure:                      { type: string, nullable: true }
            floor_area_sq_ft:            { type: integer, nullable: true }
            year_built:                  { type: integer, nullable: true }
            refurbished_year:            { type: integer, nullable: true }
            council_tax_band:            { type: string, nullable: true }
            pets_allowed:                { type: boolean, nullable: true }
            parking:                     { type: string, nullable: true }
            outdoor_space:               { type: string, nullable: true }
            furnishing:                  { type: string, nullable: true }
            deposit_amount_pence:        { type: integer, nullable: true }
            service_charge_amount_pence: { type: integer, nullable: true }
            lease_length_years:          { type: integer, nullable: true }
            available_from:              { type: string, format: date, nullable: true }
            published_at:                { type: string, format: date-time, nullable: true }
            photos:
              type: array
              items: { $ref: "#/components/schemas/Photo" }
            floor_plans:
              type: array
              items: { $ref: "#/components/schemas/FloorPlan" }
            documents:
              type: array
              items: { $ref: "#/components/schemas/Document" }
            viewing_times:
              type: array
              items: { $ref: "#/components/schemas/ViewingTime" }
            seller:
              type: object
              nullable: true
              properties:
                id:         { type: integer }
                full_name:  { type: string }
                first_name: { type: string }
    PropertyCollection:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/PropertySummary" }
        meta:  { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    Address:
      type: object
      properties:
        line_1:    { type: string }
        line_2:    { type: string, nullable: true }
        town_city: { type: string }
        county:    { type: string }
        postcode:  { type: string }
        country:   { type: string }
    Photo:
      type: object
      properties:
        id:       { type: integer }
        url:      { type: string }
        caption:  { type: string, nullable: true }
        primary:  { type: boolean }
        position: { type: integer }
    FloorPlan:
      type: object
      properties:
        id:       { type: integer }
        url:      { type: string }
        label:    { type: string, nullable: true }
        position: { type: integer }
    Document:
      type: object
      properties:
        id:           { type: integer }
        title:        { type: string }
        category:     { type: string }
        category_label:{ type: string }
        visibility:   { type: string }
        position:     { type: integer }
        download_url: { type: string }
        is_pdf:       { type: boolean }
    ViewingTime:
      type: object
      properties:
        id:         { type: integer }
        start_time: { type: string, format: date-time }
        end_time:   { type: string, format: date-time }
...
    PaginationMeta:
      type: object
      properties:
        page:        { type: integer }
        per_page:    { type: integer }
        total_pages: { type: integer }
        total_count: { type: integer }
    PaginationLinks:
      type: object
      properties:
        self: { type: string, nullable: true }
        next: { type: string, nullable: true }
        prev: { type: string, nullable: true }